requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.26.0
python-dotenv>=1.0.0
pyvis>=0.3.2
pandas>=2.0.0
numpy>=1.24.0
graphviz>=0.20.1
